Add structs VlaSegMonAct and TreePathVertexComp
VlaSegMonAct:
Let M be a monoid (i.e. a group without inversion) that acts on X.
This struct is a variable-length array storing elements in X such that
1. an element in M acting on each element in the segment [l, r]
2. retrieving the k-th element
are O(log n)-time.

TreePathVertexComp:
Let G be a group and T be a tree.
Firstly assign every vertex v a group element g_v. Then
1. reassigning a vertex,
2. given two vertices u and v, finding the chain composition g_u g_u2 ... g_v,
   where u u2 ... v is the simple path from u to v.
are O(log V)-time.